#59 Move volume generation into separate file
