+++ controllers/resource/resource_generator.go
@@ -3,15 +3,12 @@ package resource
 import (
 	"fmt"
 	"github.com/cloudogu/k8s-dogu-operator/internal"
-	"github.com/hashicorp/go-multierror"
-	"k8s.io/apimachinery/pkg/util/json"
 	"strconv"
 	"strings"
 
 	imagev1 "github.com/google/go-containerregistry/pkg/v1"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/util/intstr"
@@ -40,6 +37,9 @@ const (
 	doguPodNamespace = "POD_NAMESPACE"
 	doguPodName      = "POD_NAME"
 )
+
+// kubernetesServiceAccountKind describes a service account on kubernetes.
+const kubernetesServiceAccountKind = "k8s"
 
 // resourceGenerator generate k8s resources for a given dogu. All resources will be referenced with the dogu resource
 // as controller
@@ -109,12 +109,12 @@ func (r *resourceGenerator) CreateDoguDeployment(doguResource *k8sv1.Dogu, dogu 
 
 // GetPodTemplate returns a pod template for the given dogu.
 func (r *resourceGenerator) GetPodTemplate(doguResource *k8sv1.Dogu, dogu *core.Dogu) (*corev1.PodTemplateSpec, error) {
-	volumes, err := createVolumesForDogu(doguResource, dogu)
+	volumes, err := createVolumes(doguResource, dogu)
 	if err != nil {
 		return nil, err
 	}
 
-	volumeMounts := createVolumeMountsForDogu(doguResource, dogu)
+	volumeMounts := createVolumeMounts(doguResource, dogu)
 	envVars := []corev1.EnvVar{
 		{Name: doguPodNamespace, Value: doguResource.GetNamespace()},
 		{Name: doguPodName, ValueFrom: &corev1.EnvVarSource{
@@ -167,7 +167,7 @@ func (r *resourceGenerator) GetPodTemplate(doguResource *k8sv1.Dogu, dogu *core.
 
 func getKubernetesServiceAccount(dogu *core.Dogu) (string, bool) {
 	for _, account := range dogu.ServiceAccounts {
-		if account.Kind == string(k8sv1.KubernetesServiceAccountKind) && account.Type == k8sv1.DoguOperatorClient {
+		if account.Kind == kubernetesServiceAccountKind && account.Type == doguOperatorClient {
 			return dogu.GetSimpleName(), true
 		}
 	}
@@ -221,183 +221,6 @@ func CreateStartupProbe(dogu *core.Dogu) *corev1.Probe {
 		}
 	}
 	return nil
-}
-
-func createVolumesForDogu(doguResource *k8sv1.Dogu, dogu *core.Dogu) ([]corev1.Volume, error) {
-	nodeMasterVolume := corev1.Volume{
-		Name: nodeMasterFile,
-		VolumeSource: corev1.VolumeSource{
-			ConfigMap: &corev1.ConfigMapVolumeSource{
-				LocalObjectReference: corev1.LocalObjectReference{Name: nodeMasterFile},
-			},
-		},
-	}
-
-	mode := int32(0744)
-
-	privateVolume := corev1.Volume{
-		Name: doguResource.GetPrivateVolumeName(),
-		VolumeSource: corev1.VolumeSource{
-			Secret: &corev1.SecretVolumeSource{
-				SecretName:  doguResource.GetPrivateVolumeName(),
-				DefaultMode: &mode,
-			},
-		},
-	}
-
-	// always reserve a volume for upgrade script actions, even if the dogu has no state because upgrade scripts
-	// do not always rely on a dogu state (f. e. checks on upgradability)
-	doguReservedVolume := corev1.Volume{
-		Name: doguResource.GetReservedVolumeName(),
-		VolumeSource: corev1.VolumeSource{
-			PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{
-				ClaimName: doguResource.GetReservedPVCName(),
-			},
-		},
-	}
-
-	volumes := []corev1.Volume{
-		nodeMasterVolume,
-		privateVolume,
-		doguReservedVolume,
-	}
-
-	volumesFromDogu, err := createVolumesFromDoguVolumes(dogu.Volumes, doguResource)
-	if err != nil {
-		return nil, err
-	}
-
-	volumes = append(volumes, volumesFromDogu...)
-
-	return volumes, nil
-}
-
-func createVolumesFromDoguVolumes(doguVolumes []core.Volume, doguResource *k8sv1.Dogu) ([]corev1.Volume, error) {
-	var multiError error
-	var volumes []corev1.Volume
-	for _, doguVolume := range doguVolumes {
-		_, clientExists := doguVolume.GetClient(k8sv1.DoguOperatorClient)
-		if clientExists {
-			volume, err := createClientVolumeFromDoguVolume(doguVolume)
-			if err != nil {
-				multiError = multierror.Append(multiError, err)
-				continue
-			}
-
-			volumes = append(volumes, *volume)
-		}
-	}
-
-	dataVolumeCount := len(doguVolumes) - len(volumes)
-	if dataVolumeCount > 0 {
-		dataVolume := corev1.Volume{
-			Name: doguResource.GetDataVolumeName(),
-			VolumeSource: corev1.VolumeSource{
-				PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{
-					ClaimName: doguResource.Name,
-				},
-			},
-		}
-		volumes = append(volumes, dataVolume)
-	}
-
-	return volumes, multiError
-}
-
-func createClientVolumeFromDoguVolume(doguVolume core.Volume) (*corev1.Volume, error) {
-	client, clientExists := doguVolume.GetClient(k8sv1.DoguOperatorClient)
-	if !clientExists {
-		return nil, fmt.Errorf("dogu volume %s has no client", doguVolume.Name)
-	}
-
-	clientParams := new(k8sv1.VolumeParams)
-	err := convertGenericJsonObject(client.Params, clientParams)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read %s client params of volume %s: %w", k8sv1.DoguOperatorClient, doguVolume.Name, err)
-	}
-
-	switch clientParams.Type {
-	case k8sv1.ConfigMapParamType:
-		configMapParamContent := new(k8sv1.VolumeConfigMapContent)
-		err = convertGenericJsonObject(clientParams.Content, configMapParamContent)
-		if err != nil {
-			return nil, fmt.Errorf("failed to read %s client type content of volume %s: %w", k8sv1.ConfigMapParamType, doguVolume.Name, err)
-		}
-
-		return &corev1.Volume{
-			Name: doguVolume.Name,
-			VolumeSource: corev1.VolumeSource{
-				ConfigMap: &corev1.ConfigMapVolumeSource{
-					LocalObjectReference: corev1.LocalObjectReference{Name: configMapParamContent.Name},
-				},
-			},
-		}, nil
-	default:
-		return nil, fmt.Errorf("unsupported client param type %s in volume %s", clientParams.Type, doguVolume.Name)
-	}
-}
-
-// convertGenericJsonObject is necessary because go unmarshalls generic json objects as `map[string]interface{}`,
-// and, therefore, a type assertion is not possible. This method marshals the generic object (`map[string]interface{}`)
-// back into a string. This string is then unmarshalled back into a specific given struct.
-func convertGenericJsonObject(genericObject interface{}, targetObject interface{}) error {
-	marshalledContent, err := json.Marshal(genericObject)
-	if err != nil {
-		return err
-	}
-
-	err = json.Unmarshal(marshalledContent, targetObject)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func createVolumeMountsForDogu(doguResource *k8sv1.Dogu, dogu *core.Dogu) []corev1.VolumeMount {
-	doguVolumeMounts := []corev1.VolumeMount{
-		{
-			Name:      nodeMasterFile,
-			ReadOnly:  true,
-			MountPath: "/etc/ces/node_master",
-			SubPath:   "node_master",
-		},
-		{
-			Name:      doguResource.GetPrivateVolumeName(),
-			ReadOnly:  true,
-			MountPath: "/private",
-		},
-		{
-			Name:      doguResource.GetReservedVolumeName(),
-			ReadOnly:  false,
-			MountPath: DoguReservedPath,
-		},
-	}
-
-	for _, doguVolume := range dogu.Volumes {
-		newVolume := createVolumeMountFromDoguVolume(doguVolume, doguResource)
-		doguVolumeMounts = append(doguVolumeMounts, newVolume)
-	}
-
-	return doguVolumeMounts
-}
-
-func createVolumeMountFromDoguVolume(doguVolume core.Volume, doguResource *k8sv1.Dogu) corev1.VolumeMount {
-	_, clientExists := doguVolume.GetClient(k8sv1.DoguOperatorClient)
-	if clientExists {
-		return corev1.VolumeMount{
-			Name:      doguVolume.Name,
-			ReadOnly:  false,
-			MountPath: doguVolume.Path,
-		}
-	}
-
-	return corev1.VolumeMount{
-		Name:      doguResource.GetDataVolumeName(),
-		ReadOnly:  false,
-		MountPath: doguVolume.Path,
-		SubPath:   doguVolume.Name,
-	}
 }
 
 // GetAppLabel returns an app label which all CES resource may receive for general selection.
@@ -490,43 +313,6 @@ func (r *resourceGenerator) CreateDoguExposedServices(doguResource *k8sv1.Dogu, 
 	}
 
 	return exposedServices, nil
-}
-
-// CreateDoguPVC creates a persistent volume claim with a 5Gi storage for the given dogu.
-func (r *resourceGenerator) CreateDoguPVC(doguResource *k8sv1.Dogu) (*corev1.PersistentVolumeClaim, error) {
-	return r.createPVC(doguResource.Name, doguResource, resource.MustParse("5Gi"))
-}
-
-// CreateReservedPVC creates a persistent volume claim with a 10Mi storage for the given dogu.
-// Used for example for upgrade operations.
-func (r *resourceGenerator) CreateReservedPVC(doguResource *k8sv1.Dogu) (*corev1.PersistentVolumeClaim, error) {
-	return r.createPVC(doguResource.GetReservedPVCName(), doguResource, resource.MustParse("10Mi"))
-}
-
-func (r *resourceGenerator) createPVC(pvcName string, doguResource *k8sv1.Dogu, size resource.Quantity) (*corev1.PersistentVolumeClaim, error) {
-	pvc := &corev1.PersistentVolumeClaim{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      pvcName,
-			Namespace: doguResource.Namespace,
-			Labels:    GetAppLabel().Add(doguResource.GetDoguNameLabel()),
-		},
-	}
-
-	pvc.Spec = corev1.PersistentVolumeClaimSpec{
-		AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
-		Resources: corev1.ResourceRequirements{
-			Requests: corev1.ResourceList{
-				corev1.ResourceStorage: size,
-			},
-		},
-	}
-
-	err := ctrl.SetControllerReference(doguResource, pvc, r.scheme)
-	if err != nil {
-		return nil, wrapControllerReferenceError(err)
-	}
-
-	return pvc, nil
 }
 
 // CreateDoguSecret generates a secret with a given data map for the dogu
