+++ controllers/resource/resource_generator_test.go
@@ -2,7 +2,6 @@ package resource
 
 import (
 	_ "embed"
-	"github.com/cloudogu/k8s-dogu-operator/api/v1"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -55,7 +54,7 @@ func TestResourceGenerator_GetDoguDeployment(t *testing.T) {
 		// given
 		ldapDoguResource := readLdapDoguResource(t)
 		ldapDogu := readLdapDogu(t)
-		client, clientExists := ldapDogu.Volumes[2].GetClient(v1.DoguOperatorClient)
+		client, clientExists := ldapDogu.Volumes[2].GetClient(doguOperatorClient)
 		require.True(t, clientExists)
 		client.Params = "invalid"
 
@@ -256,74 +255,6 @@ func TestResourceGenerator_GetDoguExposedServices(t *testing.T) {
 	})
 }
 
-func TestResourceGenerator_CreateDoguPVC(t *testing.T) {
-	generator := getResourceGenerator(t)
-
-	t.Run("Return simple pvc", func(t *testing.T) {
-		// given
-		ldapDoguResource := readLdapDoguResource(t)
-
-		// when
-		actualPVC, err := generator.CreateDoguPVC(ldapDoguResource)
-
-		// then
-		require.NoError(t, err)
-		assert.Equal(t, readLdapDoguExpectedDoguPVC(t), actualPVC)
-	})
-
-	t.Run("Return error when reference owner cannot be set", func(t *testing.T) {
-		// given
-		ldapDoguResource := readLdapDoguResource(t)
-		oldMethod := ctrl.SetControllerReference
-		ctrl.SetControllerReference = func(owner, controlled metav1.Object, scheme *runtime.Scheme) error {
-			return assert.AnError
-		}
-		defer func() { ctrl.SetControllerReference = oldMethod }()
-
-		// when
-		_, err := generator.CreateDoguPVC(ldapDoguResource)
-
-		// then
-		require.Error(t, err)
-		assert.ErrorIs(t, err, assert.AnError)
-		assert.ErrorContains(t, err, "failed to set controller reference:")
-	})
-}
-
-func TestResourceGenerator_CreateReservedPVC(t *testing.T) {
-	generator := getResourceGenerator(t)
-
-	t.Run("Return simple pvc", func(t *testing.T) {
-		// given
-		ldapDoguResource := readLdapDoguResource(t)
-
-		// when
-		actualPVC, err := generator.CreateReservedPVC(ldapDoguResource)
-
-		// then
-		require.NoError(t, err)
-		assert.Equal(t, readLdapDoguExpectedReservedPVC(t), actualPVC)
-	})
-
-	t.Run("Return error when reference owner cannot be set", func(t *testing.T) {
-		// given
-		ldapDoguResource := readLdapDoguResource(t)
-		oldMethod := ctrl.SetControllerReference
-		ctrl.SetControllerReference = func(owner, controlled metav1.Object, scheme *runtime.Scheme) error {
-			return assert.AnError
-		}
-		defer func() { ctrl.SetControllerReference = oldMethod }()
-
-		// when
-		_, err := generator.CreateReservedPVC(ldapDoguResource)
-
-		// then
-		require.Error(t, err)
-		assert.ErrorIs(t, err, assert.AnError)
-		assert.ErrorContains(t, err, "failed to set controller reference:")
-	})
-}
-
 func TestResourceGenerator_GetDoguSecret(t *testing.T) {
 	generator := getResourceGenerator(t)
 
@@ -370,65 +301,5 @@ func Test_createLivenessProbe(t *testing.T) {
 
 		// then
 		require.Nil(t, actual)
-	})
-}
-
-func Test_createClientVolumeFromDoguVolume(t *testing.T) {
-	t.Run("should fail due to missing client", func(t *testing.T) {
-		// given
-		doguVolume := core.Volume{
-			Name:    "my-volume",
-			Clients: []core.VolumeClient{},
-		}
-
-		// when
-		_, err := createClientVolumeFromDoguVolume(doguVolume)
-
-		// then
-		require.Error(t, err)
-		assert.ErrorContains(t, err, "dogu volume my-volume has no client")
-	})
-	t.Run("should fail due to invalid config map type content", func(t *testing.T) {
-		// given
-		doguVolume := core.Volume{
-			Name: "my-volume",
-			Clients: []core.VolumeClient{
-				{
-					Name: "k8s-dogu-operator",
-					Params: v1.VolumeParams{
-						Type:    "configmap",
-						Content: "invalid",
-					},
-				},
-			},
-		}
-
-		// when
-		_, err := createClientVolumeFromDoguVolume(doguVolume)
-
-		// then
-		require.Error(t, err)
-		assert.ErrorContains(t, err, "failed to read configmap client type content of volume my-volume")
-	})
-	t.Run("should fail due to unsupported client param type", func(t *testing.T) {
-		// given
-		doguVolume := core.Volume{
-			Name: "my-volume",
-			Clients: []core.VolumeClient{
-				{
-					Name: "k8s-dogu-operator",
-					Params: v1.VolumeParams{
-						Type: "invalid",
-					},
-				},
-			},
-		}
-
-		// when
-		_, err := createClientVolumeFromDoguVolume(doguVolume)
-
-		// then
-		require.Error(t, err)
-		assert.ErrorContains(t, err, "unsupported client param type invalid in volume my-volume")
 	})
 }
