+++ controllers/serviceaccount/serviceaccount_creator.go
@@ -24,6 +24,9 @@ import (
 	"github.com/pkg/errors"
 )
 
+// doguKind describes a service account on a dogu.
+const doguKind = "dogu"
+
 // creator is the unit to handle the creation of service accounts
 type creator struct {
 	client      client.Client
@@ -48,7 +51,7 @@ func (c *creator) CreateAll(ctx context.Context, dogu *core.Dogu) error {
 	logger := log.FromContext(ctx)
 
 	for _, serviceAccount := range dogu.ServiceAccounts {
-		if serviceAccount.Kind != "" && serviceAccount.Kind != string(v1.DoguServiceAccountKind) {
+		if serviceAccount.Kind != "" && serviceAccount.Kind != string(doguKind) {
 			continue
 		}
