+++ controllers/serviceaccount/serviceaccount_remover.go
@@ -3,7 +3,6 @@ package serviceaccount
 import (
 	"context"
 	"fmt"
-	v12 "github.com/cloudogu/k8s-dogu-operator/api/v1"
 	"github.com/cloudogu/k8s-dogu-operator/internal"
 
 	"github.com/cloudogu/cesapp-lib/core"
@@ -44,7 +43,7 @@ func (r *remover) RemoveAll(ctx context.Context, dogu *core.Dogu) error {
 	var allProblems error
 
 	for _, serviceAccount := range dogu.ServiceAccounts {
-		if serviceAccount.Kind != "" && serviceAccount.Kind != string(v12.DoguServiceAccountKind) {
+		if serviceAccount.Kind != "" && serviceAccount.Kind != string(doguKind) {
 			continue
 		}
